quarter-2-openai-agents-sdk/class-08/handsoff/pyproject.toml
[project]
name = "handsoff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
]
